+++ gazebo/gui/model/ModelData_TEST.cc
@@ -32,16 +32,16 @@ void ModelData_TEST::Clone()
 
   // Create the main window.
   gazebo::gui::MainWindow *mainWindow = new gazebo::gui::MainWindow();
-  QVERIFY(mainWindow != NULL);
+  QVERIFY(mainWindow != nullptr);
   mainWindow->Load();
   mainWindow->Init();
   mainWindow->show();
 
   // Get the user camera and scene
   gazebo::rendering::UserCameraPtr cam = gazebo::gui::get_active_camera();
-  QVERIFY(cam != NULL);
+  QVERIFY(cam != nullptr);
   gazebo::rendering::ScenePtr scene = cam->GetScene();
-  QVERIFY(scene != NULL);
+  QVERIFY(scene != nullptr);
 
   this->ProcessEventsAndDraw(mainWindow);
 
@@ -56,7 +56,7 @@ void ModelData_TEST::Clone()
   link->Load(msgs::LinkToSDF(model.link(0)));
   rendering::VisualPtr linkVis(new rendering::Visual("model::box_link",
       scene->WorldVisual()));
-  link->linkVisual = linkVis;
+  link->SetLinkVisual(linkVis);
 
   // add a visual
   rendering::VisualPtr vis(
@@ -76,15 +76,15 @@ void ModelData_TEST::Clone()
   QCOMPARE(cloneLink->GetName(), cloneLinkName);
   QCOMPARE(cloneLink->Scale(), ignition::math::Vector3d::One);
   QCOMPARE(cloneLink->Pose(), ignition::math::Pose3d::Zero);
-  QVERIFY(cloneLink->linkVisual != NULL);
-  QCOMPARE(cloneLink->linkVisual->GetName(), "model::" + cloneLinkName);
+  QVERIFY(cloneLink->LinkVisual() != nullptr);
+  QCOMPARE(cloneLink->LinkVisual()->GetName(), "model::" + cloneLinkName);
   QVERIFY(cloneLink->Scale() == ignition::math::Vector3d::One);
 
   // verify clone link visual
   QCOMPARE(cloneLink->visuals.size(), link->visuals.size());
   QVERIFY(cloneLink->visuals.size() == 1u);
   rendering::VisualPtr cloneVis = cloneLink->visuals.begin()->first;
-  QVERIFY(cloneVis != NULL);
+  QVERIFY(cloneVis != nullptr);
   QCOMPARE(cloneVis->GetName(), "model::" + cloneLinkName + "::visual");
   QCOMPARE(cloneVis->GetGeometryType(), std::string("box"));
   QCOMPARE(cloneVis->GetGeometrySize(), size);
@@ -93,7 +93,7 @@ void ModelData_TEST::Clone()
   QCOMPARE(cloneLink->collisions.size(), link->collisions.size());
   QVERIFY(cloneLink->collisions.size() == 1u);
   rendering::VisualPtr cloneCol = cloneLink->collisions.begin()->first;
-  QVERIFY(cloneCol != NULL);
+  QVERIFY(cloneCol != nullptr);
   QCOMPARE(cloneCol->GetName(), "model::" + cloneLinkName + "::collision");
   QCOMPARE(cloneCol->GetGeometryType(), std::string("box"));
   QCOMPARE(cloneCol->GetGeometrySize(), size);
@@ -145,7 +145,7 @@ void ModelData_TEST::Clone()
 
   mainWindow->close();
   delete mainWindow;
-  mainWindow = NULL;
+  mainWindow = nullptr;
 }
 
 /////////////////////////////////////////////////
@@ -158,16 +158,16 @@ void ModelData_TEST::LinkScale()
 
   // Create the main window.
   gazebo::gui::MainWindow *mainWindow = new gazebo::gui::MainWindow();
-  QVERIFY(mainWindow != NULL);
+  QVERIFY(mainWindow != nullptr);
   mainWindow->Load();
   mainWindow->Init();
   mainWindow->show();
 
   // Get the user camera and scene
   gazebo::rendering::UserCameraPtr cam = gazebo::gui::get_active_camera();
-  QVERIFY(cam != NULL);
+  QVERIFY(cam != nullptr);
   gazebo::rendering::ScenePtr scene = cam->GetScene();
-  QVERIFY(scene != NULL);
+  QVERIFY(scene != nullptr);
 
   this->ProcessEventsAndDraw(mainWindow);
 
@@ -183,7 +183,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddBoxLink(model, mass, size);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("box_link", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -332,7 +332,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddCylinderLink(model, mass, radius, length);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("cylinder_link", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -486,7 +486,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddSphereLink(model, mass, radius);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("sphere_link", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -634,7 +634,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddSphereLink(model, mass, radius);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("sphere_link2", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -734,7 +734,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddSphereLink(model, mass, radius);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("sphere_link3", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -811,7 +811,7 @@ void ModelData_TEST::LinkScale()
 
   mainWindow->close();
   delete mainWindow;
-  mainWindow = NULL;
+  mainWindow = nullptr;
 }
 
 // Generate a main function for the test